# backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Finance Backend API
  version: 0.1.0
servers:
  - url: http://localhost:8080/api
paths:
  /auth/register:
    post:
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: Registered
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authenticated
  /assets:
    get:
      summary: List assets
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Assets
  /transactions:
    get:
      summary: List transactions
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Transactions
    post:
      summary: Create transaction
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Transaction created
  /dividends:
    get:
      summary: List dividends
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Dividends
  /portfolio:
    get:
      summary: Get current portfolio snapshot
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Portfolio snapshot
  /portfolio/summary:
    get:
      summary: Get portfolio summary
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Summary
  /portfolio/performance:
    get:
      summary: Get portfolio performance series
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Performance
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    RegisterRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
    LoginRequest:
      type: object
      required: [email, password]
      additionalProperties: false
      properties:
        email:
          type: string
          format: email
        password:
          type: string